// frontend_geek/src/app/enums/position.ts
//  La position représente les différentes positions que peut occuper un travailleur 
export enum Position {
  CEO = 'CEO',
  Director = 'Director',
  ProjectManager = 'ProjectManager',
  Developer = 'Developer',
  SystemAnalyst = 'SystemAnalyst',
  ITSupport = 'ITSupport',
  NetworkEngineer = 'NetworkEngineer',
}